app/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
boto3==1.34.40
gunicorn==21.2.0
python-dotenv==1.0.1
